Move database management from Dashboard to Settings page and simplify Dashboard to placeholder home page
- Move all database CRUD functionality (create, update, delete, activate) from Dashboard to Settings page
- Move database list table, active database info card, and all modals to Settings
- Replace Dashboard content with simple home page header using Home icon
- Add navigate prop to Settings component for collection navigation
- Update App.jsx to pass navigate prop to Settings
- Maintain all existing database
